requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
asyncio==4.0.0
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
banks==2.4.0
beautifulsoup4==4.14.3
cbor2==5.8.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
constantly==23.10.4
croniter==6.2.2
cryptography==46.0.3
daphne==4.2.1
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
Django==5.2.10
django-environ==0.12.0
django-filter==25.2
django-jsonform==2.23.2
django-picklefield==3.4.0
django-q2==1.9.0
djangorestframework==3.16.1
et_xmlfile==2.0.0
filetype==1.2.0
frozenlist==1.8.0
fsspec==2026.2.0
greenlet==3.3.1
griffe==2.0.0
griffecli==2.0.0
griffelib==2.0.0
h11==0.16.0
html-to-markdown==2.28.2
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
iniconfig==2.3.0
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index==0.14.14
llama-index-cli==0.5.3
llama-index-core==0.14.14
llama-index-embeddings-ollama==0.8.6
llama-index-embeddings-openai==0.5.1
llama-index-indices-managed-llama-cloud==0.9.4
llama-index-instrumentation==0.4.2
llama-index-llms-ollama==0.9.1
llama-index-llms-openai==0.6.18
llama-index-readers-file==0.5.6
llama-index-readers-llama-parse==0.5.1
llama-index-workflows==2.14.2
llama-parse==0.6.54
MarkupSafe==3.0.3
marshmallow==3.26.2
msgpack==1.1.2
multidict==6.7.1
mypy_extensions==1.1.0
mysqlclient==2.2.7
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.2
numpy==2.4.2
ollama==0.6.1
openai==2.21.0
openpyxl==3.1.5
packaging==26.0
pandas==2.3.3
pillow==12.1.0
platformdirs==4.9.1
playwright==1.57.0
pluggy==1.6.0
propcache==0.4.1
py-ubjson==0.16.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.0
Pygments==2.19.2
pyOpenSSL==25.3.0
pypdf==6.7.0
pytest==9.0.2
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pyttsx3==2.99
pytz==2025.2
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.46
sqlparse==0.5.5
striprtf==0.0.26
supervisor==4.3.0
tenacity==9.1.4
tiktoken==0.12.0
tqdm==4.67.3
Twisted==25.5.0
txaio==25.12.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.11.0
urllib3==2.6.3
whitenoise==6.11.0
wrapt==1.17.3
yarl==1.22.0
zope.interface==8.2

chromadb~=1.5.5
huggingface_hub~=1.8.0